+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2023.2.9]
+
+### `Project Import` enhancements
+- Register `bootstrap/gensrc` and other `src` directories as source directories with respect of readonly state [#718](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/718)
+
 ## [2023.2.8]
 
 ### Features